requirements.txt
langchain-unstructured
langchain-community
unstructured-client
unstructured
"unstructured[ocr]"
"unstructured[pdf]" 
python-magic
paddleocr
pytesseract
pi_heif
